python/requirements.txt
hydra-core==1.1
omegaconf==2.1.0
networkx==2.6.2
scipy==1.4.1
numpy==1.21.0
nptyping==1.4.3
pandas==1.3.3
seaborn==0.11.2
hydra-joblib-launcher==1.1.5
